Add support to use one repository per feature file - This will add more flexibility when testing. Now each feature file uses a different GitHub repo to avoid flaky tests due to data modified by other Scenario from other feature file. - Stop using only one repo per whole suite(Data Management). - There's a bug on pytest side on teardown method when using parallelism. Commented line with bug reference is provided.

+++ conftest.py
@@ -1,6 +1,12 @@
+import os
+import re
+
 import pytest
 from os import getenv
 from typing import Generator
+
+from _pytest.nodes import Item
+
 from api.Client import APIClient
 from config.Config import Config
 
@@ -34,34 +40,58 @@ def request_context():
     return {}
 
 
-@pytest.fixture(scope="session", autouse=True)
-def create_test_repository(api_client, config
+@pytest.fixture(scope="function", autouse=True)
+def create_test_repository(api_client, config, request,
                            ) -> Generator[None, None, None]:
+    data_tag = None
+    for temp_marker in request.node.own_markers:
+        marker = temp_marker.name
+        if marker.startswith("data:"):
+            data_tag = marker.split(":")[1].strip()
+            config.add_repo(data_tag)
+            break
+
+    if not data_tag:
+        raise ValueError("Feature file requires @data tag, but none was found")
+
+    # create of a GitHub repository for each feature file
     # Check if the repository already exists
     check_repo = api_client.make_request(
         method="GET",
-        endpoint=f"/repos/{config.GITHUB_USER}/{config.GITHUB_REPO}"
+        endpoint=f"/repos/{config.GITHUB_USER}/{data_tag}"
     )
-
     if check_repo['status'] == 200:
-        print(f"Repository {config.GITHUB_REPO} already exists. Skipping creation.")
+        print(f"Repository {data_tag} already exists. Skipping creation.")
     else:
         # Before all
         new_repo = api_client.make_request(
             method="POST",
             endpoint="/user/repos",
-            payload={"name": f'{config.GITHUB_REPO}'}
+            payload={"name": f'{data_tag}'}
         )
-        assert new_repo['status'] == 201, f"Failed to create repository. Status: {new_repo.status}, Response: {new_repo.text}"
-        print(f"Repository {config.GITHUB_REPO} created successfully.")
+        assert new_repo[
+                   'status'] == 201, f"Failed to create repository. Status: {new_repo.status}, Response: {new_repo.text}"
+        print(f"Repository {data_tag} created successfully.")
 
+
+@pytest.fixture(scope="session", autouse=True)
+def delete_test_repositories(api_client, config
+                             ) -> Generator[None, None, None]:
     yield
-    # After all
     worker_id = getenv("PYTEST_XDIST_WORKER", "master")
-    if worker_id == "master":
-        deleted_repo = api_client.make_request(
-            method="DELETE",
-            endpoint=f"/repos/{config.GITHUB_USER}/{config.GITHUB_REPO}"
-        )
-        assert deleted_repo['status'] == 204, f"Failed to delete repository. Status: {deleted_repo.status}"
-        print(f"Repository {config.GITHUB_REPO} deleted successfully.")
+    if worker_id == "master" and config.get_repositories():
+        # Bug: https://github.com/pytest-dev/pytest-xdist/issues/1029
+        # no worker master when running tests in parallel
+        # https://pytest-xdist.readthedocs.io/en/latest/how-to.html#making-session-scoped-fixtures-execute-only-once
+        for repo_name in config.get_repositories():
+            deleted_repo = api_client.make_request(
+                method="DELETE",
+                endpoint=f"/repos/{config.GITHUB_USER}/{repo_name}"
+            )
+            if deleted_repo['status'] == 204:
+                print(f"Repository {repo_name} deleted successfully.")
+        # Delete JSON file with GitHub repositories
+        if config.get_repositories():
+            path_to_file = config.SHARED_DATA_FILE
+            if os.path.exists(path_to_file):
+                os.remove(path_to_file)
